+++ .editorconfig
@@ -22,4 +22,3 @@ indent_style = space
 [{package.json,.travis.yml}]
 indent_size = 2
 indent_style = space
-
