+++ course_selection_app/data/areas.pkl
