+++ course_selection_app/data/current_semester.pkl
