+++ course_selection_app/model/constants.py
