+++ course_selection_app/static/.eslintrc.json
