+++ course_selection_app/static/css/index.css
