+++ course_selection_app/static/js/calendar.js
