+++ course_selection_app/static/js/class_table.js
