+++ course_selection_app/static/js/utility.js
