+++ course_selection_app/templates/index.html
@@ -55,8 +55,8 @@
           href="https://fonts.googleapis.com/css?family=Questrial|Roboto+Condensed|Nanum+Gothic">
 
     <!-- Project javascript and CSS -->
-    <script type="module" src="../static/js/script_base.js"></script>
-    <link rel="stylesheet" href="../static/css/style_base.css">
+    <script type="module" src="../static/js/index.js"></script>
+    <link rel="stylesheet" href="../static/css/index.css">
 </head>
 <!---------------------------------------------------------------------------->
