+++ run.py
@@ -1,0 +1,4 @@
+from course_selection_app.application import app
+
+if __name__ == "__main__":
+    app.run(threaded=True)
