--- static/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "Dependencies",
-  "version": "1.0.0",
-  "description": "Run npm install to install all needed denpendencies.",
-  "dependencies": {
-    "jquery": "^3.3.1",
-    "bootstrap": "^4.1.1",
-    "popper.js": "^1.14.3",
-    "fullcalendar": "^3.9.0",
-    "jquery-confirm": "^3.3.2",
-    "bootstrap-select": "^1.13.1",
-    "@fortawesome/fontawesome-free": "^5.1.0"
-  },
-  "devDependencies": {
-    "eslint": "^5.1.0",
-    "eslint-plugin-node": "^7.0.1",
-    "eslint-plugin-jsdoc": "^3.7.1",
-    "eslint-plugin-import": "^2.13.0",
-    "eslint-plugin-promise": "^3.8.0",
-    "eslint-plugin-standard": "^3.1.0",
-    "eslint-config-standard": "^12.0.0-alpha.0"
-  },
-  "author": "Weiqi",
-  "license": "ISC"
-}
